refactor: BaseTest 패키지 변경
