+++ src/test/java/com/debatetimer/repository/parliamentary/ParliamentaryTableRepositoryTest.java
@@ -3,11 +3,11 @@ package com.debatetimer.repository.parliamentary;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.debatetimer.BaseRepositoryTest;
 import com.debatetimer.domain.member.Member;
 import com.debatetimer.domain.parliamentary.ParliamentaryTable;
 import com.debatetimer.exception.custom.DTClientErrorException;
 import com.debatetimer.exception.errorcode.ClientErrorCode;
+import com.debatetimer.repository.BaseRepositoryTest;
 import java.util.List;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
