+++ src/test/java/com/debatetimer/repository/parliamentary/ParliamentaryTimeBoxRepositoryTest.java
@@ -2,10 +2,10 @@ package com.debatetimer.repository.parliamentary;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.debatetimer.BaseRepositoryTest;
 import com.debatetimer.domain.member.Member;
 import com.debatetimer.domain.parliamentary.ParliamentaryTable;
 import com.debatetimer.domain.parliamentary.ParliamentaryTimeBox;
+import com.debatetimer.repository.BaseRepositoryTest;
 import java.util.List;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
